✨ MLIR | Add `PlacementPass` (#1232)
## Description

This pull request splits the routing pass into a placement and routing
pass.

### Changes 

The new placement pass replaces `alloc`s and `dealloc`s with static
qubits, generates and applies the initial placement, and handles the
extension of results and arguments of `scf::For` and `scf::If`.

As a consequence the routing pass becomes much more simpler. Since the
placement pass doesn't have to worry about SWAPs, the
`HardwareIndexPool` is replaced by a simple `std::deque<std::size_t>`.
Following the same argument, the awkward `checkZeroUse` check can be
avoided entirely.

To reduce code duplication, I've also moved the `Layout` (previously
`LayoutState`) to a separate `Layout.h` file and use it throughout all
transpilation passes. Likewise for constants and the `isEntryPoint`
helper function there is now a `Common.h` file.

### Initial Placement (Help appreciated 🙌🏼)

I'm still very much unsatisfied with the initial placement solution.
Even though the interface is quite clear (_"return program <-> hardware
mapping"_) the current solution still feels too disintegrated to allow
LightSABRE-like logic. Maybe it's possible to give the `InitialPlacer` a
copy of the `func.func` (or something along the lines of that).

Ultimately, in the near future, we probably have to implement an initial
placement strategy that requires the underlying computation (such as the
subgraph isomorphism problem) to address these issues properly.

---

Moreover: How do we want to specify, or record, the initial placement
vector for equivalence checking down the line?


## Checklist:

<!---
This checklist serves as a reminder of a couple of things that ensure
your pull request will be merged swiftly.
-->

- [x] The pull request only contains commits that are focused and
relevant to this change.
- [x] I have added appropriate tests that cover the new/changed
functionality.
- [x] I have updated the documentation to reflect these changes.
- [x] I have added entries to the changelog for any noteworthy
additions, changes, fixes, or removals.
- [x] I have added migration instructions to the upgrade guide (if
needed).
- [x] The changes follow the project's style guidelines and introduce no
new warnings.
- [x] The changes are fully tested and pass the CI checks.
- [x] I have reviewed my own code changes.

---------

Signed-off-by: burgholzer <burgholzer@me.com>
Co-authored-by: burgholzer <burgholzer@me.com>