✨ DD-based Simulator QDMI device (#1287)
## Description

This PR adds a new QDMI device to the available devices in MQT Core:
A classical simulator device based on the same Decision Diagram (DD)
approach that also powers MQT DDSIM.
The device supports OpenQASM2 and OpenQASM3 as well as a wide variety of
gates.
It supports weak simulation (=sampling) and strong simulation (=state
vector simulation) based on the specified number of shots (`0` shots
indicate that strong simulation is desired).
The sampling mode fully supports dynamic quantum circuits with
mid-circuit measurements, resets, and conditional gates.

This marks the first MQT QDMI device that implements the full job
interface.

At the moment this is still an early draft that just provides the first
draft of the corresponding device. It lacks proper embedding as well as
testing within MQT Core.

## Checklist:

<!---
This checklist serves as a reminder of a couple of things that ensure
your pull request will be merged swiftly.
-->

- [x] The pull request only contains commits that are focused and
relevant to this change.
- [x] I have added appropriate tests that cover the new/changed
functionality.
- [x] I have updated the documentation to reflect these changes.
- [x] I have added entries to the changelog for any noteworthy
additions, changes, fixes, or removals.
- [x] I have added migration instructions to the upgrade guide (if
needed).
- [x] The changes follow the project's style guidelines and introduce no
new warnings.
- [x] The changes are fully tested and pass the CI checks.
- [x] I have reviewed my own code changes.

---------

Signed-off-by: burgholzer <burgholzer@me.com>