🚨 Enable Astral's type checker `ty` (#1333)
## Description

This PR enables running Astral's newest project `ty`, a type checker
meant to replace mypy and similar tools.
For the moment, this is added _in addition_ to mypy and does not yet
replace it entirely.
Once `ty` moves out of alpha status officially, we will completely
remove mypy. That is also when we will update the documentation and the
development guide.

## Checklist:

<!---
This checklist serves as a reminder of a couple of things that ensure
your pull request will be merged swiftly.
-->

- [x] The pull request only contains commits that are focused and
relevant to this change.
- [x] I have added appropriate tests that cover the new/changed
functionality.
- [x] I have updated the documentation to reflect these changes.
- [x] I have added entries to the changelog for any noteworthy
additions, changes, fixes, or removals.
- [x] I have added migration instructions to the upgrade guide (if
needed).
- [x] The changes follow the project's style guidelines and introduce no
new warnings.
- [x] The changes are fully tested and pass the CI checks.
- [x] I have reviewed my own code changes.

---------

Signed-off-by: burgholzer <burgholzer@me.com>