✨Implement Unitized Routing Passes (#1301)
## Description

This pull request implements an "unit"-ized approach to route
quantum-classical programs. Furthermore, it solves the issue that the
def-use and textual (program) operation order is not _that_ compatible.

### Unit

An Unit divides a quantum-classical program into routable sections:

```
┌─────┐
│unit |
└─────┘
---------------- divider (`scf.for`)
┃ ┌─────┐
┃ |unit | // loop-body
┃ └─────┘ <-- `restore=true` 
┌─────┐
│unit |
└─────┘
---------------- divider (`scf.if`)
┃ ┌─────┐
┃ |unit | // then-body
┃ └─────┘<-- `restore=true` 
┣
┃ ┌─────┐
┃ │unit | // else-body
┃ └─────┘<-- `restore=true` 
┌─────┐
│unit |
└─────┘
```

Each unit is routed separately. This PR introduces two modes of
traversal for such units:
- `SequentialUnit`: Traverses the ops in the unit sequentially (Naive
Routing)
- `LayeredUnit`: Traverses the ops layer-by-layer (A* Routing) 

## Changes
- Split naive (`--route-naive-sc`) from A* routing pass
(`--route-astar-sc`): This makes sense from a logical as well as a pass
option perspective. The pass options for the A* routing pass belong to
the A* routing pass. If we were to implement another method, let's say
SABRE, its options would again reside in a separate pass.
- Implement `SequentialUnit` and `LayeredUnit`.
- Refactor A* and Naive routing passes to use the respective unit
implementation.
- Fix SSA Dominance Issues by implementing the reordering strategy
discussed below.
- Use `SequentialUnit` also for verification 


## Checklist:

<!---
This checklist serves as a reminder of a couple of things that ensure
your pull request will be merged swiftly.
-->

- [x] The pull request only contains commits that are focused and
relevant to this change.
- [x] I have added appropriate tests that cover the new/changed
functionality.
- [x] I have updated the documentation to reflect these changes.
- [x] I have added entries to the changelog for any noteworthy
additions, changes, fixes, or removals.
- [x] I have added migration instructions to the upgrade guide (if
needed).
- [x] The changes follow the project's style guidelines and introduce no
new warnings.
- [x] The changes are fully tested and pass the CI checks.
- [x] I have reviewed my own code changes.

[^1]: Switching the custom stack for recursive function calls can
probably also be applied to the placement and verification pass.
Something for a future PR.

---------

Signed-off-by: matthias <matthias@bereumann.com>
Co-authored-by: pre-commit-ci[bot] <66853113+pre-commit-ci[bot]@users.noreply.github.com>
Co-authored-by: Lukas Burgholzer <burgholzer@me.com>