:bug: Handling Conditional Branches from QuantumComputation to MLIR (#1378)
## Description

This pull request fixes a bug that arises if one wants to cast a
QuantumComputation to MLIR which has been imported from an OpenQASM
program.
Currently, if one imports a QASM file with conditional branching and
casts the file to MLIR, the contents of the if- and else-branches are
ignored. This is due to the fact that the qasm Importer creates
CompoundOperations from the branch contents.
ImportQuantumComputation.cpp meanwhile can only handle Operations. See
added test, which fails under the current conditions due to empty if-
and else branches.
I have changed ImportQuantumComputation.cpp to handle both
CompoundOperations and Operations inside if- and else-blocks.

## Checklist:

<!---
This checklist serves as a reminder of a couple of things that ensure
your pull request will be merged swiftly.
-->

- [x] The pull request only contains commits that are focused and
relevant to this change.
- [x] I have added appropriate tests that cover the new/changed
functionality.
- [x] I have updated the documentation to reflect these changes.
- [x] I have added entries to the changelog for any noteworthy
additions, changes, fixes, or removals.
- [ ] I have added migration instructions to the upgrade guide (if
needed).
- [x] The changes follow the project's style guidelines and introduce no
new warnings.
- [ ] The changes are fully tested and pass the CI checks.
- [x] I have reviewed my own code changes.